file

+++ view/Web.js
@@ -1,6 +1,6 @@
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import Constants from "expo-constants";
-import { KeyboardAvoidingView, AsyncStorage, View } from "react-native";
+import { KeyboardAvoidingView, AsyncStorage, View, Alert } from "react-native";
 import { WebView } from "react-native-webview";
 import { useDispatch, useSelector } from "react-redux";
 import { removeAllSelectedAsset } from "../modules/selectedAssetList";
@@ -8,11 +8,15 @@ import { showLoading, hideLoading } from "../modules/loading";
 import Loading from "../components/GalleryComp/Loading";
 import { setAlbumInfo } from "../modules/album";
 import { ProgressBar, Colors } from "react-native-paper";
-import { hideProgress } from "../modules/progress";
 
 const Web = ({ navigation }) => {
+  const [visible, setVisible] = useState(false);
   const dispatch = useDispatch();
   const click = async (token) => {
+    if (progress < 1 && progress > 0) {
+      Alert.alert("파일을 올리는 중입니다.");
+      return;
+    }
     dispatch(showLoading());
     dispatch(removeAllSelectedAsset());
     dispatch(setAlbumInfo({}));
@@ -21,9 +25,12 @@ const Web = ({ navigation }) => {
   };
   const progress = useSelector((s) => s.progress);
   useEffect(() => {
-    console.log(progress);
-    dispatch(hideProgress());
-  }, []);
+    if (progress < 1 && progress > 0) {
+      setVisible(true);
+    } else {
+      setVisible(false);
+    }
+  }, [progress]);
   return (
     <>
       <Loading />
@@ -35,7 +42,7 @@ const Web = ({ navigation }) => {
         behavior="height"
       >
         <WebView
-          source={{ uri: "http://172.30.1.52:3000" }}
+          source={{ uri: "http://172.30.1.6:3000" }}
           onMessage={(event) => click(event.nativeEvent.data)}
           style={{ marginTop: Constants.statusBarHeight }}
           renderLoading={() => {
@@ -43,11 +50,14 @@ const Web = ({ navigation }) => {
           }}
           onLoad={() => dispatch(hideLoading())}
         />
-        <ProgressBar
-          progress={progress.percent}
-          color={Colors.blue700}
-          visible={progress.visible}
-        />
+
+        {visible && (
+          <ProgressBar
+            progress={progress}
+            color={Colors.blue700}
+            style={{ height: 10 }}
+          />
+        )}
       </KeyboardAvoidingView>
     </>
   );
